requirements.txt
PyJWT==2.4.0
PyYAML==6.0
fastapi==0.75.0
pydantic==1.9.0
tortoise-orm==0.19.0
uvicorn==0.17.6
requests==2.27.1
gunicorn==20.1.0
aiomysql==0.0.22
dependency_injector==4.39.1
aerich==0.6.3
passlib==1.7.4
aioredis==2.0.1